short code

+++ src/main/java/org/example/Main.java
@@ -8,7 +8,7 @@ public class Main {
 
     public static void main(String[] args) {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         System.out.println("Данные о продажах предприятия по месяцам:");
         System.out.println(Arrays.toString(sales));
         System.out.println("Cумма всех продаж");
