+++ src/test/java/StatsServiceTest.java
@@ -3,13 +3,13 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 public class StatsServiceTest {
-    int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+    long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
     @Test
     public void allSales() { //сумма
         StatsService service = new StatsService();
         int expected = 180;
-        int actual = service.sumSales(sales);
+        long actual = service.sumSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -17,7 +17,7 @@ public class StatsServiceTest {
     public void averageSalesAmount() { //средняя сумма
         StatsService service = new StatsService();
         int expected = 15;
-        int actual = service.averageSumSales(sales);
+        long actual = service.averageSumSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -25,7 +25,7 @@ public class StatsServiceTest {
     public void numberMaxSales() { //№макс
         StatsService service = new StatsService();
         int expected = 8;
-        int actual = service.maxSales(sales);
+        long actual = service.maxSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -33,7 +33,7 @@ public class StatsServiceTest {
     public void numberMinSales() { //№мин
         StatsService service = new StatsService();
         int expected = 9;
-        int actual = service.minSales(sales);
+        long actual = service.minSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -41,7 +41,7 @@ public class StatsServiceTest {
     public void averageMinSalesAmount() { //ниже среднего
         StatsService service = new StatsService();
         int expected = 5;
-        int actual = service.averageMinSales(sales);
+        long actual = service.averageMinSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -49,7 +49,7 @@ public class StatsServiceTest {
     public void averageMaxSalesAmount() { //выше среднего
         StatsService service = new StatsService();
         int expected = 5;
-        int actual = service.averageMaxSales(sales);
+        long actual = service.averageMaxSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 }
